Don't perform edit menu operations if not allowed ...
- add sibling (if root node is selected)
- delete node (if root node is selected)
- undo (if undo is not available)
- redo (if redo is not available)

+++ src/components/menus/EditMenu.ts
@@ -132,15 +132,27 @@ function onAddChildButtonClick() {
 }
 
 function onAddSiblingButtonClick() {
-    state.ui.setCurrentModal('addSibling');
+    const rootNodeId = state.doc.getRootNodeId();
+    const selectedNodeId = state.doc.getSelectedNodeId();
+
+    if (selectedNodeId !== rootNodeId) {
+        state.ui.setCurrentModal('addSibling');
+    }
 }
 
 function onDeleteNodeButtonClick() {
-    state.doc.deleteNode(state.doc.getSelectedNodeId());
+    const rootNodeId = state.doc.getRootNodeId();
+    const selectedNodeId = state.doc.getSelectedNodeId();
+
+    if (selectedNodeId !== rootNodeId) {
+        state.doc.deleteNode(state.doc.getSelectedNodeId());
+    }
 }
 
 function onRedoButtonClick() {
-    state.doc.redo();
+    if (state.doc.getRedoIsAvailable()) {
+        state.doc.redo();
+    }
 }
 
 function onReplaceNodeContentsButtonClick() {
@@ -148,7 +160,9 @@ function onReplaceNodeContentsButtonClick() {
 }
 
 function onUndoButtonButtonClick() {
-    state.doc.undo();
+    if (state.doc.getUndoIsAvailable()) {
+        state.doc.undo();
+    }
 }
 
 export default EditMenu;
